fix tests, handle flag params in socket:call
