+++ packages/cli/src/commands/socket/call.ts
@@ -1,6 +1,7 @@
 import {flags} from '@oclif/command'
 import format from 'chalk'
 import inquirer from 'inquirer'
+import yargs from 'yargs-parser'
 
 import Command, {Socket} from '../../base_command'
 import {echon, printCode, printSourceCode} from '../../utils/print-tools'
@@ -10,13 +11,13 @@ export default class SocketEndpointCall extends Command {
   static flags = {
     'body-only': flags.boolean()
   }
+  static strict = false
   static args = [{
     name: 'fullEndpointName',
     required: true,
     description: 'full endpoint name in format: <socket_name>/<endpoint_name>'
   }]
-
-  static validateValue(value) {
+  static validateValue(value: any) {
     if (!value) {
       return 'We need this!'
     }
@@ -24,15 +25,8 @@ export default class SocketEndpointCall extends Command {
     return true
   }
 
-  context: any
-  session: any
-  Socket: any
-  socketList: any
-  mainSpinner: any
-  cmd: any
-  localSockets: any
-
   promptParamQuestion(params, param) {
+    const flags = yargs(this.argv)
     const description = params[param].description || ''
     const paramType = params[param].type
     this.echo(4)(`- ${param} ${format.dim(`(${paramType})`)} ${description}`)
@@ -40,7 +34,8 @@ export default class SocketEndpointCall extends Command {
       name: param,
       message: this.p(2)(`Type in value for "${format.green(param)}" parameter`),
       default: params[param].example,
-      validate: value => SocketEndpointCall.validateValue(value)
+      validate: value => SocketEndpointCall.validateValue(value),
+      when: flags[param] === undefined
     }
     return question
   }
@@ -62,12 +57,14 @@ export default class SocketEndpointCall extends Command {
   }
 
   listParams(endpointObj) {
+    const flags = yargs(this.argv)
     const inputs = endpointObj.metadata.inputs
     const params = inputs ? inputs.properties || {} : {}
+    const flagValues = {}
     const paramsCount = Object.keys(params).length
     const questions = []
 
-    if (!paramsCount) return questions
+    if (!paramsCount) return {questions, flagValues}
 
     this.echo()
     echon(4)(`You can pass ${format.cyan(paramsCount.toString())} `)
@@ -76,18 +73,18 @@ export default class SocketEndpointCall extends Command {
 
     Object.keys(params).forEach(param => {
       questions.push(this.promptParamQuestion(params, param))
+      flagValues[param] = flags[param]
     })
     this.echo()
 
-    return questions
+    return {questions, flagValues}
   }
 
   async run() {
-    await this.session.isAuthenticated()
-    await this.session.hasProject()
+    this.session.isAuthenticated()
+    this.session.hasProject()
 
-    const {args} = this.parse(SocketEndpointCall)
-    const {flags} = this.parse(SocketEndpointCall)
+    const {args, flags} = this.parse(SocketEndpointCall)
 
     try {
       const bodyOnly = flags['body-only']
@@ -96,11 +93,12 @@ export default class SocketEndpointCall extends Command {
       const endpointObj = await socket.getEndpoint(endpointName)
 
       if (endpointObj && endpointObj.existRemotely) {
-        const askQuestions = this.listParams(endpointObj)
+        const {questions, flagValues} = this.listParams(endpointObj)
         let config = {}
-        if (askQuestions.length > 0) {
-          config = await inquirer.prompt(askQuestions) || {}
+        if (questions.length > 0) {
+          config = await inquirer.prompt(questions) || {}
         }
+        config = {...flagValues, ...config}
         try {
           const res = await endpointObj.call(config)
           this.formatResponse(res, bodyOnly)
